pyproject.toml
[project]
name = "cwm-cdn-api"
version = "0.1.0"
description = "CWM CDN API"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.2.1",
    "fastapi[standard]>=0.116.1",
    "idna>=3.10",
    "orjson>=3.11.1",
    "prometheus-client>=0.23.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "retry>=0.9.2",
    "uvicorn>=0.35.0",
    "uvicorn-worker>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
cwm-cdn-api = "cwm_cdn_api.cli:main"

[project.optional-dependencies]
load-test = [
    "locust>=2.43.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
